Fix: val is a int type in set_elem()

+++ matrix.h
@@ -11,6 +11,6 @@ Matrix* create_matrix_from_file(FILE* file);
 Matrix* create_matrix(int row, int col);
 void free_matrix(Matrix* matrix);
 double get_elem(Matrix* matrix, int row, int col);
-void set_elem(Matrix* matrix, int row, int col, int val);
+void set_elem(Matrix* matrix, int row, int col, double val);
 
 #endif //_MATRIX_H_
